Fix render texture fix
No seriously how did the original author not notice that the texture is fine,
just the alpha is fucked? I mean FUCK, you know how much time I lost to this
shit? FUCK YOU. I already solved that with the old cunt exporter but thought
it's a different problem now because this is the shiny new exporter but no. No,
it's the same fucking problem. Bloody stupid.